requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
anthropic>=0.18.0
openai>=1.0.0
requests>=2.31.0
